This is an android application that give the user the meaning of the word when the mobile camera is shown to the word .
It is an extended image processing application that helps the user to
explore the meaning of a new word which is in the printed format. The
application captures the text area using the camera of the smart phone
and process the captured image to extract the text which is in the
editable form. This process is done by optical character recognition
technique (OCR) and carried on by cloud service. The extracted text
content containing the required word is provided to the user in a
editable text field for confirming precision. Once when the word is
confirmed, using the dictionary API the definition for the word is
obtained. The image processing done in the cloud uses efficient OCR
techniques to convert the printed text into digitized and editable text
format thus by reducing the framework on the user side and making the
application more efficient and powerful with less requirements.